requirements.txt
annotated-types==0.7.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
cryptography==43.0.0
et-xmlfile==1.1.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
httplib2==0.22.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.18.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pypdf==4.3.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sympy==1.13.1
torch==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
xgboost==2.1.1
XlsxWriter==3.2.0
